move verifiererror traits to their own file
